Make package versions compatible with nodejs v10
When trying to `npm run build`, gulp crashes with a weird assertion
error described on SO here:

https://stackoverflow.com/questions/50620673/

and in the gulp issue tracker here:

https://github.com/gulpjs/gulp/issues/2162

Turns out the issue has to do with the 'natives' package.  After some
investigation, triggering a dependency rewrite with `npm install
gulp@3.9` seems to do the trick with the least resistance.